requirements.txt
requests==2.25.0
beautifulsoup4==4.9.3
Pillow==8.0.1
